requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0
python-multipart==0.0.6
sqlalchemy>=2.0.0
passlib>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
jinja2>=3.1.0
authlib>=1.3.0
httpx>=0.27.0
python-dotenv>=1.0.0
itsdangerous>=2.1.0

# MongoDB dependencies
motor==3.3.2
pymongo==4.6.1
beanie==1.24.0

